requirements.txt
# What: Production dependencies only
# Why: Clean production deployments without dev tools
# When: Used by Docker, CI/CD, production servers
torch>=1.12.0
torchvision>=0.13.0
grpcio>=1.47.0
grpcio-tools>=1.47.0
protobuf>=3.19.0
kubernetes>=24.2.0
redis>=4.3.0
psycopg2-binary>=2.9.0
prometheus-client>=0.14.0
pyyaml>=6.0
numpy>=1.21.0